Completes support for specifying tag by slaveId on input
The motivation for this is to allow specifying the tag and sensor from
deconstructing the (default) topic (which is ${mac}/${slaveId}/${sensorType}).

This isn't currently needed by this node, but allows identifying a tag or
sensor to be actuated from, say, an MQTT topic with the same basis as the
message topic output by these nodes.